refactor(e2e): share jsonl dir reading via playwright-vscode-ext
readJsonlFiles/parseJsonlLines replace the spec-local readSessionFiles/parseLines in
spanRedaction.desktop.spec.ts, so the new spec stops adding a fourth copy of the
readdir -> filter .jsonl -> JSON.parse-per-line helper.
